+++ src/test/java/nextstep/ladder/domain/LadderTest.java
@@ -9,7 +9,7 @@ import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 class LadderTest {
     @Test
     void lines_생성() {
-        Ladder ladder = Ladder.of(List.of("a", "b", "c", "d", "e"), 5);
+        Ladder ladder = LadderFactory.createLadder(List.of("a", "b", "c", "d", "e"), 5);
         assertThat(ladder.getLines().size()).isEqualTo(5);
     }
 }
